update_7_7_23_teacherinformation

+++ database/migrations/2023_05_18_170928_create_rooms_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('Room_Name', 20);
             $table->integer('Quantity');
             $table->text('Note')->nullable();
-            $table->smallInteger('Status');
+            $table->integer('Status');
         });
     }
