+++ database/migrations/2023_05_18_175213_create_batch_of_good_details_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->integer('Batch_Of_Goods_ID');
             $table->integer('Quantity');
             $table->decimal('Unit_Price', 10, 2);
-            $table->smallInteger('Status');
+            $table->integer('Status');
         });
     }
