+++ database/migrations/2023_05_20_033539_create_requires_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->integer('Device_ID');
             $table->smallInteger('Require_Type');
             $table->text('Description');
-            $table->smallInteger('Status');
+            $table->integer('Status');
             $table->timestamps();
         });
     }
